Move delete button (Case:47398)
There was feedback that it's too easy to accidentally delete a question.  Since
a confirm dialog or an undo facility would take more time to implement and only
partially address the problem, we're moving the delete button.

There's nowhere on the toolbar to move it, so we're moving it to the question
properties box.

Minimal change, would require substantial re-architecting to do anything
prettier.

Tested on recent Chrome and FF.