Add ability to duplicate a question (Case:48046)
Adds the option to duplicate a question, either maintaining the original itext
id values ("link itext"), or setting all questions in the duplicate to auto
itext id, thereby creating new itext nodes for everything ("copy itext").

Question IDs are renamed according to a foo -> foo-1, foo-2... pattern.

When you duplicate an entire tree of questions, all logic expressions in the
duplicate are updated so any relative references to questions within the tree
refer to the appropriate question in the duplicate, and any references to
questions outside the tree maintain their original reference.

There are only two things this doesn't do that I would want it to:

1. Linking of Itext such that when you edit the display text of one question,
any other questions using the same itext immediately have their name in the
tree updated (and more importantly, their display text input value).  This
actually works correctly if you save the form and refresh the page, but that's
not very useful.  I can probably figure this out.

2. Insert duplicate question after source question instead of at the end of
source question's parent's tree.  Originally had it working this way, but ended
up abandoning that due to a lot of difficulty with certain edge cases due to
the way we handle relative insert position.