Fix data node issues with duplicate question (Case:48589)
There were many issues adding data nodes and it seemed like new ones popped up
when I solved other ones.

For instance, after I had perfectly working code that added the duplicate
question after instead of at the end, I couldn't add data nodes at all, leading
to the discovery that in our use of the jstree types plugin, we weren't
specifying datanode as a valid child of the root node, and we had somehow been
hacking around that to put nodes in the UI tree.

1. I had known before that we had two different points of truth for what types
could be children of what other types, but when I had it shoved in my face like
this and realized that it lost me a significant amount of time, I decided to
refactor the code and remove the redundant checkMoveOp() function and redundant
handler code.

I also wanted to do this so it would be easy to solve the issue of trying to
add a data node while a leaf node is selected and the issue of adding a
question while a select item is selected (Case:37184) solvable in a generic,
readable way in only one place in the code.

2. The setCurrentlySelectedMugType() and getCurrentlySelectedMugType() methods
of formdesigner.controller were used to maintain a local copy of the UI's state
and had to be called whenever you did jstree.select_node or
jstree.deselect_all, which I did extensively when duplicating a question, and I
found this to be extremely confusing and error-prone, so I changed those to
mutate and inspect the actual UI's state and removed redundant calls to those
functions.

Most anything else is just the minimal refactoring necessary to make other
things not break with the above changes, mainly 1.