remove constraint itext id if constraint message is deleted (Case:54143)
If the user clears a constraint message (or hint) text input, clear the
constraint itext id advanced field.

There is a bug when there are multiple language inputs and a user clears
one of them.  The itext id will be deleted even if the other language
itext input still has a value.  But if the user either re-enters a new
value for the language they cleared or clears the other language then it
will end up in the correct state.  Amelia said this is better than the
current behavior.