don't eat odkx:intent tag in head (FB:55436)
This adds the ability to specify tag names of first-level children of
<head> to store on parsing and re-output when serializing the XML.  The
nodes are simply stored as a list of XML DOM objects on the form and
serialized using a simple cross-browser compatible XML node to string
function.

I also cleaned up some of the selectors to be more explicit (for
example, only looking for title in the children of head, not the entire
xml body, and changed some selectors to be more compatible across
browser implementations (jQuery doesn't abstract away browser
differences in support for finding namespaced node names).