discard references to itext items with no value
formhub forms have itext for everything by default, even if it's just an
empty value.  This determines which itext ids have no actual value for
all translations and then discards references to those itext ids while
parsing the control block.

This also adds a special case for automatically converting formhub's
'eng' to our 'en' language code.  This is necessary for the above to be
useful because otherwise if opts.langs = ['en', 'hin'] and it's the
first time viewing a formhub form, there will be 'en', 'hin', and 'eng'
in the form.