fix choice default value broken after refactoring
mug.typeSlug is no longer defined manually in the mugtype
definition.  It's now just the mugType class name, and is set after
mug creation in a wrapper.  (We eventually, if it's ever worthwhile,
want to switch to prototypes and instanceof and get rid of typeSlug
altogether -- probably the only real justification would be using
prototypes to make long form loads faster by not creating duplicates of
every function on every mugtype object.)

This caused a problem in this one instance where during mug creation
typeSlug was referenced.  Moving the value definition to a better place
fixes.