copy linked itext on id change instead of changing id in all questions
This is a minimal change that adds a reference count to itext items.
When editing question properties and changing an itext id for an itext
item that's referenced by more than one question, the itext item will be
duplicated and the duplicate given the new id, while the existing item
and references will remain unchanged.

This allows you to more easily fix mistakes made when normalizing your
itext.

Tested with normal, constraint, and hint itext ids, with > 2 linked
itexts, and also renaming the original on one of the unchanged
questions, and verifying everything in the XML.