remove more unnecessary indirection in widget definitions
all sections simply return a list of property paths now.  Properties are
shown or hidden based on presence of the property definition + the
visibility attribute of the property definition.

If the visibility attribute is a path to another property, then the
property will be visible if that property is visible (this is used for
itext widgets -- the actual display toggling is handled by the id
widgets, though).