fix constraint id not allowed error for groups
somewhere in the refactoring, where we started to use the property
definitions more than they were actually being used before, it caused an
issue where you would set the constraint itext id on all mugs as you
parsed the bindlist, and then when you parsed the control tree and saw
that a mug was a group/repeat/field list (which don't allow constraint
messages), it would still set the constraint itext id on that mug, which
caused an error to show up.

This fixes that by ignoring properties where the spec says they aren't
allowed when calling mugElement.setAttr().  This also paves the way for
eventually less worrying about what question types can be changed to
what other question types, because it would all be handled here.