fix unintended breakage
We need a way to initially set the datatype, even if it should be
immutable otherwise!  This solution is kind of ugly, but I've got a
better solution in the works.

(It works as is for tagName on the controlElement because that is
currently set without going through the setAttr() interface, in the mug
constructors.)