retain raw attributes in data and bind nodes
since the parsing creates and destroys new mugs after seeing a mug's
data, bind, and control elements, in that order, it was not correctly
copying the dataElement._rawAttributes property, so raw attributes on
data nodes were lost.

This fixes that, and also makes arbitrary attributes on bind elements
get retained as well.