remove unnecessary MugElement indirection
The whole dataElement, bindElement, and controlElement separation was a
source of unnecessary complexity in the spec system, widgets, and
elsewhere.

It led to lots of boilerplate like this:

  if (mug.dataElement) {
      // do stuff with mug.dataElement.someProperty
  }

Or

  if (!mug.dataElement) {
      // expects a control-only node.  Not very explicit.
  }

Now, there is only one property container, accessed like:

  mug.p.someProperty

And all accessing and setting of properties on it triggers model change
handlers.  This change basically isolates that code in one function.

Additionally, this enabled some changes to the mug type and spec
definition system that make it yet more declarative.

Fixes:

 fixes #100: shore up model change encapsulation

Also fixed at some point:

 fixes #113: copying choices doesn't generate a unique ID after the first copy
 fixes #15: Ability to change question type to/from label