Make plugin method call order deterministic
Previously plugin methods were appended to the global $.vellum object based on when the plugin was loaded, which was non-deterministic. Now plugin methods are attached to each Vellum instance at when it is instantiated. This has the additional benefit of reducing the overhead of plugin method calls because they do not need to check if they are enabled for the instance on which they are being called.

Fixes an intermittent bug in the interaction between ignore-but-retain and javaRosa plugins. Ignore-but-retain’s loadXML method could be called after javaRosa’s, which caused ignored itext nodes to be picked up by javaRosa.