Fix tests failing only in my local environment
Also remove filter from test script now that it's not needed.

There seems to be a timing bug with asynchronously loading the file-uploader component, which subsequently caused the destroy method to fail, breaking tests. This was only observed in my local environment. Other developers and Travis did not see the problem. The bug was uncovered by https://github.com/dimagi/Vellum/pull/303, which seems entirely unrelated.