Fix cursor hiding after insert expression bubble
There are still strange editing behaviors related to whitespace after widgets (bubbles) in CKEditor. The most annoying to me is on backspace or left-arrow to remove or move within spaces following a bubble (sometimes the widget is selected instead of the desired action). Also selecting a different mug and then returning causes the ZWS chars to be removed or converted to regular spaces, which can cause a mildly inconsistent editing experience.