Move escaped hashtag logic into richText
Nothing outside of the richText module needs escaped hashtags, and richText only needs escaped hashtags in invalid xpath expressions where the xpath parser cannot be used to find them. Escaped hashtags are now only used in expressions tagged with #invalid/xpath, and are always persisted in  "vellum:" attributes, which are (or should be) ignored when rich text is disabled.