mug.supportsRichText() -> mug.form.richText
I'm a little conflicted by this change because it looks like not OOP best practice, but I think the simplicity of always looking in the same place to see if rich text is enabled is better than having the extra level of indirection. The OOP argument would say we're planning for (or at least suggesting the possibility of) some time in the future when a mug can support rich text in some scenario where it does not simply check if rich text is enabled for the form--seems unlikely.