Fix CKEditor skipping many spaces
`pre-wrap` caused a string of normal spaces (char 32), rather than alternating space/`$nbsp;`, to be entered (at least in Chrome), which was a problem because normal spaces are skipped by CKEditor when navigating toward a widget with arrow keys or backspace/del.

This fix was tested in Firefox and Chrome.