Fix save form with new language and missing itext
TypeError: Cannot read property 'replace' of undefined
 at fixEmptyTags (src/xml.js:167:26)
 at Object.query (src/xml.js:36:20)
 at writeValue (src/javaRosa/plugin.js:416:27)

When a new language has been added to the form, markdown itext for questions that are not selected during a form edit session may not be not initialized. This caused an error on write XML (save). Note that the first question is usually selected on initial form load, so this bug did not occur on forms having a single question.