Restored UI behavior for existing media area
This renames hqm-status to hqm-errors, becuase it's only used to display errors
(in multi-file versions of the uploader, it would do more). It intentionally removes the
"Cancel Upload" button shown when there are errors as a simplification. Closing and re-opening the model also
allows the user to uplaod a new file.